requirements.txt
asgiref==3.3.1
Django==3.1.7
django-scheduler==0.9.3
django-suit==0.2.28
icalendar==4.0.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
